Updating datalogging to use floats rather than doubles

+++ include/cassie_controllers/standing_control_qp.hpp
@@ -29,7 +29,7 @@ public:
     void reset();
     bool reconfigure();
     bool isReadyToTransition() {return this->memory.readyToTransition;}
-    void getDebug(VectorXd &dbg);
+    void getDebug(VectorXf &dbg);
 
 private:
     /**
